refactor: appType 드롭다운 컴포넌트화
